Moved interceptSocket to IOService
THis makes the ConnectionManager more consistent since it can focus
on higher level connection management instead of providing all kinds
of lower level socket functions.

And a good example of 'tell; don't ask'.
